[drivers/fs] Always use register_mtddriver() to register the MTD device.
We have adjusted the registration method for MTD ,
replacing the previous approach using ftl_initialize() and
bchdev_register() with register_mtddriver().

After registering with register_mtddriver(), the behavior
of the device will vary depending on the access method:
1. When accessing the device node via the open() interface,
   the MTD device will be automatically converted to a character
   device, provided that BCH support is enabled.
   During this process, both FTL (Flash Translation Layer) and
   BCH wrappers will be implicitly added.
2. When calling find_blockdriver() followed by open_blockdriver(),
   the MTD device will be treated as a block device, with only
   the FTL wrapper automatically applied.

For these reasons, the legacy registration methods ftl_initialize()
and bchdev_register() are no longer required and will be used as
internal interface.

Signed-off-by: jingfei <jingfei@xiaomi.com>